make /v2 the default, expose DEIS_API_VERSION
Deis v2 is currently in development. For users wanting to use this
with the v1 API, they can set DEIS_API_VERSION.

NOTE: This is a very basic feature flag, and as Deis v2 progresses
API endpoints are likely to be backwards-incompatible. At some point
a better integration would be to write separate v1 and v2 API client
libraries, or to completely deprecate v1 compatibility.